% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_breathlessness_scale}
\alias{add_breathlessness_scale}
\title{add_breathlessness_scale}
\usage{
add_breathlessness_scale(dat)
}
\arguments{
\item{dat}{A dataset with relevant variables}
}
\value{
`dat` with `out_mmrc_scale` added
}
\description{
Add breathlessness scale variable
}
