% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_d28_corrections}
\alias{add_d28_corrections}
\title{add_d28_corrections}
\usage{
add_d28_corrections(dat)
}
\arguments{
\item{dat}{Day 28 dataset}
}
\value{
dat with corrections
}
\description{
Make corrections to the day 28 dataset
}
