% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_database_corrections}
\alias{add_database_corrections}
\title{add_database_corrections}
\usage{
add_database_corrections(dat)
}
\arguments{
\item{dat}{Dataset containing relevant variables}
}
\value{
`dat` but with fields corrected
}
\description{
Some corrections to the data base are required, but cannot be made
to the database directly. These corrections are listed in
"ASCOT_ADAPT_DatabaseCorrectionsErrors.xlsx" and are here applied
manually to the data extracts.
}
