% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_days_alive_and_free}
\alias{add_days_alive_and_free}
\title{add_days_alive_and_free}
\usage{
add_days_alive_and_free(dat)
}
\arguments{
\item{dat}{A dataset with relevant variables}
}
\value{
`dat` with `out_dafh` for days alive and free of hospital
}
\description{
Add days alive and free of hospital
A death is counted as 0 days
}
