% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_days_alive_and_free_ventilation}
\alias{add_days_alive_and_free_ventilation}
\title{add_days_alive_and_free_ventilation}
\usage{
add_days_alive_and_free_ventilation(dat)
}
\arguments{
\item{dat}{A dataset with relevant variables}
}
\value{
`dat` with `out_dafv` for days alive and free of ventilation
}
\description{
Add days alive and free of ventilation
A death is counted as 0 days
}
