% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{add_derived_covariates}
\alias{add_derived_covariates}
\title{Add derived covariates}
\usage{
add_derived_covariates(dat)
}
\arguments{
\item{dat}{A dataset}
}
\value{
A dataset with added covariates
}
\description{
Adds a collection of manually coded covariates (e.g. converted "Yes"/"No" data to 1/0 data etc)
}
