% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{add_epoch_term}
\alias{add_epoch_term}
\title{Add epoch term to dataset}
\usage{
add_epoch_term(dat)
}
\arguments{
\item{dat}{A dataset with `RandDate`}
}
\value{
A dataset with epoch variables
}
\description{
Adds an epoch term to the dataset where each epoch
is a 4 week window from the most recent randomisation.
If any epoch as less than 5 participants, then
it is combined with the most recent adjacent epoch.
}
