% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_outcome_corrections}
\alias{add_outcome_corrections}
\title{add_outcome_corrections}
\usage{
add_outcome_corrections(dat)
}
\arguments{
\item{dat}{Dataset containing relevant variables}
}
\value{
`dat` but with relevant outcome fields corrected
}
\description{
Additional follow-up with sites was undertaken for participants
randomsied to the antiviral domain who had missing primary outcome data
due to unknown day 28 status.

The following corrections are made as per email correspondence with study
coordinators who contacted sites for further information.
}
