% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_oxygen}
\alias{add_oxygen}
\title{add_oxygen}
\usage{
add_oxygen(dat)
}
\arguments{
\item{dat}{Dataset containing relevant variables}
}
\value{
Returns dat with new variables.
}
\description{
Add requirement of oxygen at baseline
}
