% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_primary_outcome_components}
\alias{add_primary_outcome_components}
\title{add_primary_outcome}
\usage{
add_primary_outcome_components(dat)
}
\arguments{
\item{dat}{Dataset with one record per participant with their outcomes}
}
\value{
Returns `dat` but with primary outcome fields appended.
}
\description{
Add the derived primary outcome for each participant.
}
