% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{add_region_site_groups}
\alias{add_region_site_groups}
\title{Add region and site groupings to dataset}
\usage{
add_region_site_groups(dat)
}
\arguments{
\item{dat}{A dataset with `Country` and `Location` variables}
}
\value{
A dataset with ctry and site variables.
}
\description{
Adds region groups and site nested within regions to dataset.
Sites with less than 5 participants are combined into
an "other" sites group within region.
}
