% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_shortness_of_breath}
\alias{add_shortness_of_breath}
\title{add_shortness_of_breath}
\usage{
add_shortness_of_breath(dat)
}
\arguments{
\item{dat}{A dataset with relevant variables}
}
\value{
`dat` with `out_sob` added
}
\description{
Add a shortness of breath variable
}
