% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_supplemental_oxygen}
\alias{add_supplemental_oxygen}
\title{add_supplemental_oxygen}
\usage{
add_supplemental_oxygen(dat)
}
\arguments{
\item{dat}{Dataset containing relevant variables}
}
\value{
Returns dat with new variables.
}
\description{
Add requirement of supplemental oxygen at baseline
Per email from Steve:
For the baseline status, we have that initial question regarding ‘Was the patient on room air for any of the preceding 24 hours?’
If no – they needed supplemental oxygen.
If yes – the next Q can be used – if O2 saturation <94%, then we code as needing supplemental O2. If ≥94%, then we code as NOT needing supplemental oxygen.
}
