% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_time_to_recovery}
\alias{add_time_to_recovery}
\title{add_time_to_recovery}
\usage{
add_time_to_recovery(dat)
}
\arguments{
\item{dat}{A dataset with relevant variables}
}
\value{
`dat` with `out_rec`, and `out_ttr` added
}
\description{
Add time to recovery variable
}
