% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_v5_activation_date}
\alias{add_v5_activation_date}
\title{add_v5_activation_date}
\usage{
add_v5_activation_date(dat)
}
\arguments{
\item{dat}{A dataset}
}
\value{
dat with site data added
}
\description{
Add site activation data
}
