% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{add_withdrawn_followup}
\alias{add_withdrawn_followup}
\title{add_withdrawn_followup}
\usage{
add_withdrawn_followup(dat)
}
\arguments{
\item{dat}{Dataset containing relevant withdrawal variables}
}
\value{
Returns dat with a new indicator variable for whether
participant withdrew from follow-up.
}
\description{
Add withdrawal flag to participant records.
}
