% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{compile_cmdstanr_mod}
\alias{compile_cmdstanr_mod}
\title{compile_cmdstanr_mod
Compile provided cumulative logistic Stan model for cmdstanr.}
\usage{
compile_cmdstanr_mod(mod_name, ...)
}
\arguments{
\item{mod_name}{The model name}
}
\value{
A `CmdStanModel` object
}
\description{
compile_cmdstanr_mod
Compile provided cumulative logistic Stan model for cmdstanr.
}
