% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{create_and_save_derived_data}
\alias{create_and_save_derived_data}
\title{create_and_save_derived_data}
\usage{
create_and_save_derived_data(dir1, dir2)
}
\arguments{
\item{dir1}{Directory for raw data}

\item{dir2}{Directory for derived data}
}
\description{
create_and_save_derived_data
}
