% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{create_fulldata_add_daily}
\alias{create_fulldata_add_daily}
\title{create_fulldata_add_daily}
\usage{
create_fulldata_add_daily(dat)
}
\arguments{
\item{dat}{Full dataset from `create_fulldata_no_daily`}

\item{daily}{Daily dataset}
}
\description{
Adds daily data to the dataset
}
