% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{create_fulldata_no_daily}
\alias{create_fulldata_no_daily}
\title{create_fulldata_no_daily()}
\usage{
create_fulldata_no_daily()
}
\value{
Full dataset without daily records
}
\description{
Basically joins all the datasets together, except for the daily records.
Perform additional formatting here, in sub-steps for each raw data table.
}
