% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{creatinine_clearance}
\alias{creatinine_clearance}
\title{creatinine_clearance}
\usage{
creatinine_clearance(sex, age, weight, creatinine)
}
\arguments{
\item{sex}{Either "Male" or "Female"}

\item{age}{Age in years}

\item{weight}{Weight in kg}

\item{creatinine}{Serum creatinine in umol/L}
}
\value{
Creatinine clearance rate in mL/min.
}
\description{
Serum creatinine clearance by Cockcroft-Gault formula
https://www.eviq.org.au/clinical-resources/eviq-calculators/3200-creatinine-clearance-calculator
}
