% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{decision_quantity_summary_table}
\alias{decision_quantity_summary_table}
\title{Summarise decision quantities}
\usage{
decision_quantity_summary_table(OR, format = "html", fut_val = 1.1)
}
\arguments{
\item{OR}{RVs}

\item{format}{Table format}

\item{fut_val}{Futility reference value}
}
\value{
A tibble
}
\description{
Summarise decision quantities
}
