% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{filter_acs_itt}
\alias{filter_acs_itt}
\title{ACS-ITT set}
\usage{
filter_acs_itt(dat)
}
\arguments{
\item{dat}{Dataset}
}
\value{
Filtered dataset
}
\description{
Filter the full dataset down to the ACS-ITT,
that is, those who were enrolled and did not
withdraw from follow-up and were randomised to
intervention in domain C.
}
