% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{filter_concurrent_intermediate}
\alias{filter_concurrent_intermediate}
\title{filter_concurrent_intermediate}
\usage{
filter_concurrent_intermediate(dat)
}
\arguments{
\item{dat}{Dataset}
}
\value{
Filtered dataset
}
\description{
Filter data to those in the intermediate-dose concurrent dataset
}
