% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{filter_concurrent_std_aspirin}
\alias{filter_concurrent_std_aspirin}
\title{filter_concurrent_std_aspirin}
\usage{
filter_concurrent_std_aspirin(dat)
}
\arguments{
\item{dat}{Dataset}
}
\value{
Filtered dataset
}
\description{
Filter data to those in the low-dose with aspirin concurrent dataset
}
