% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{filter_fas_itt}
\alias{filter_fas_itt}
\title{filter_fas_itt}
\usage{
filter_fas_itt(dat)
}
\arguments{
\item{dat}{Dataset}
}
\value{
Filtered dataset
}
\description{
Filter the full dataset down to the FAS-ITT,
that is, those who were enrolled and did not
withdraw from follow-up.
}
