% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findfirst}
\alias{findfirst}
\title{findfirst}
\usage{
findfirst(x, v = NA)
}
\arguments{
\item{x}{Logical vector}

\item{v}{Default if no TRUE elements}
}
\value{
First TRUE element of x
}
\description{
findfirst
}
