% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_primary_model}
\alias{fit_primary_model}
\title{Fit primary model}
\usage{
fit_primary_model(
  dat = NULL,
  model = NULL,
  vars = c("inelgc3", "agegte60", "ctry"),
  beta_sd_int = 2.5,
  beta_sd_var = c(10, 2.5, 1, 1),
  beta_sd_trt = 1,
  ctr = contr.equalprior,
  seed = 32915,
  ...
)
}
\arguments{
\item{dat}{Dataset (e.g. output from `make_fas_itt_set`)}

\item{model}{Stan model}

\item{vars}{Covariates to include}

\item{beta_sd_int}{Prior SD for intercept}

\item{beta_sd_var}{Prior SD for vars}

\item{beta_sd_trt}{Prior SD for treatments}

\item{ctr}{Contrast for treatment design}

\item{seed}{Seed}

\item{...}{Other arguments}
}
\description{
Creates Stan dataset, samples from the model, and generates summaries
}
