% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{format_baseline_data}
\alias{format_baseline_data}
\title{format_baseline_data}
\usage{
format_baseline_data(bas)
}
\arguments{
\item{bas}{Baseline data}
}
\value{
Formatted data
}
\description{
format_baseline_data
}
