% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{format_consent_data}
\alias{format_consent_data}
\title{format_consent_data}
\usage{
format_consent_data(con)
}
\arguments{
\item{con}{Consent data}
}
\value{
Formatted data
}
\description{
format_consent_data
}
