% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{format_daily_data}
\alias{format_daily_data}
\title{format_daily_data}
\usage{
format_daily_data(dd)
}
\arguments{
\item{dd}{Raw daily data extract}
}
\value{
Formatted daily data
}
\description{
Apply formatting to the raw daily data extract.
}
