% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{format_discharge_data}
\alias{format_discharge_data}
\title{format_discharge_data}
\usage{
format_discharge_data(dis)
}
\arguments{
\item{dis}{Discharge data}
}
\value{
Formatted data
}
\description{
format_discharge_data
}
