% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{format_eligibility_data}
\alias{format_eligibility_data}
\title{format_eligibility_data}
\usage{
format_eligibility_data(el)
}
\arguments{
\item{el}{Eligibility data}
}
\value{
Formatted data
}
\description{
Format eligibility data
}
