% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{format_enrolled_data}
\alias{format_enrolled_data}
\title{format_enrolled_data}
\usage{
format_enrolled_data(enr)
}
\arguments{
\item{enr}{Enrolled data}
}
\value{
Formatted data
}
\description{
format_enrolled_data
}
