% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{format_withdrawal_data}
\alias{format_withdrawal_data}
\title{format_withdrawal_data}
\usage{
format_withdrawal_data(wth)
}
\arguments{
\item{wth}{Withdrawal data}
}
\value{
Formatted data
}
\description{
format_withdrawal_data
}
