% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_baseline_comorbidities}
\alias{generate_baseline_comorbidities}
\title{generate_baseline_comorbidities}
\usage{
generate_baseline_comorbidities(dat)
}
\arguments{
\item{dat}{Dataset with baseline variables}
}
\value{
A tibble giving the summary
}
\description{
Generates baseline co-morbidity summary across all participants
}
