% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_baseline_comorbidities_by}
\alias{generate_baseline_comorbidities_by}
\title{generate_baseline_comorbidities_by}
\usage{
generate_baseline_comorbidities_by(dat, grpvar = NULL)
}
\arguments{
\item{dat}{The dataset}

\item{grpvar}{The grouping variable}
}
\value{
A tibble giving the summary
}
\description{
Generates baseline co-morbidity summary across grouping variable
}
