% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_baseline_comorbidities_table}
\alias{generate_baseline_comorbidities_table}
\title{generate_baseline_comorbidities_table}
\usage{
generate_baseline_comorbidities_table(dat, format = "html")
}
\arguments{
\item{dat}{The dataset}

\item{closed}{If TRUE, one table per domain, otherwise just aggregated table}
}
\value{
A list of tibbles giving the summary tables
}
\description{
Generates baseline co-morbidity summary tables for each domain
}
