% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_baseline_demographics}
\alias{generate_baseline_demographics}
\title{generate_baseline_demographics}
\usage{
generate_baseline_demographics(dat)
}
\arguments{
\item{data}{The dataset. assumed to have baseline data included}
}
\description{
Generate baseline demographics summary overall
}
