% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_baseline_demographics_by}
\alias{generate_baseline_demographics_by}
\title{generate_baseline_demographics_by}
\usage{
generate_baseline_demographics_by(dat, grpvar = NULL)
}
\arguments{
\item{grpvar}{The grouping variable}

\item{data}{The dataset. assumed to have baseline data included}
}
\description{
Generate baseline demographics summary by a grouping variable
}
