% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_baseline_demographics_table}
\alias{generate_baseline_demographics_table}
\title{generate_baseline_demographics_table}
\usage{
generate_baseline_demographics_table(dat, format = "html")
}
\arguments{
\item{data}{The dataset. assumed to have baseline data included
If FALSE produce one table per domain.}
}
\value{
Either a single table or a list of tables, one for each domain.
}
\description{
Generate baseline demographics tables
}
