% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_baseline_prognostics}
\alias{generate_baseline_prognostics}
\title{Baseline prognostics overall}
\usage{
generate_baseline_prognostics(dat)
}
\arguments{
\item{dat}{The dataset}
}
\value{
A tibble giving the summary
}
\description{
Generates baseline prognostics summary overall
}
