% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_baseline_prognostics_table}
\alias{generate_baseline_prognostics_table}
\title{Baseline prognostics table}
\usage{
generate_baseline_prognostics_table(dat, format = "html")
}
\arguments{
\item{dat}{The dataset}

\item{closed}{If TRUE, generates summary only overall, if FALSE generates by domain}
}
\value{
A tibble giving the summary
}
\description{
Generates baseline prognostics summary tables
}
