% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_completeness.R
\name{generate_completeness_data}
\alias{generate_completeness_data}
\title{generate_completeness_data}
\usage{
generate_completeness_data(ddat)
}
\arguments{
\item{ddat}{The current daily dataset}
}
\value{
A tibble of data completeness
}
\description{
Generate completeness for the daily data of participants who
are 28 days past their date of randomisation.

Each participant should have:
- one baseline record
- one discharge record
- one daily record per day up until the date of their discharge
- one day 28 record
}
