% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_discharge_drugs}
\alias{generate_discharge_drugs}
\title{Drugs used during hospital stay overall}
\usage{
generate_discharge_drugs(dat)
}
\arguments{
\item{dat}{The dataset}
}
\value{
A tibble giving the summary
}
\description{
Generates overall summary of drugs used during hospital stay,
as recorded on discharge
}
