% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_discharge_drugs_by}
\alias{generate_discharge_drugs_by}
\title{Drugs used during hospital stay by group}
\usage{
generate_discharge_drugs_by(dat, grpvar = NULL)
}
\arguments{
\item{dat}{The dataset}

\item{grpvar}{The grouping variable}
}
\value{
A tibble giving the summary
}
\description{
Generates summary of drugs used during hospital stay,
as recorded on discharge, by grouping variable
}
