% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{generate_table_1_data}
\alias{generate_table_1_data}
\title{generate_table_1_data}
\usage{
generate_table_1_data(dat, grp = NULL)
}
\arguments{
\item{dat}{A dataset}
}
\description{
generate_table_1_data
}
