% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mappings.R
\name{intervention_labels}
\alias{intervention_labels}
\title{intervention_labels}
\usage{
intervention_labels()
}
\value{
A list of intervention labels with explicit breaks
}
\description{
Get intervention labels for all domains
}
