% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mappings.R
\name{intervention_labels2}
\alias{intervention_labels2}
\title{intervention_labels2}
\usage{
intervention_labels2()
}
\value{
A list of intervention labels without breaks
}
\description{
Get intervention labels for all domains
}
