% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mappings.R
\name{intervention_labels_short}
\alias{intervention_labels_short}
\title{intervention_labels_short}
\usage{
intervention_labels_short()
}
\value{
A list of short intervention labels
}
\description{
Get intervention labels for all domains
}
