% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mappings.R
\name{intervention_labels_short_break}
\alias{intervention_labels_short_break}
\title{intervention_labels_short_break}
\usage{
intervention_labels_short_break()
}
\value{
A list of short intervention labels with breaks
}
\description{
Get intervention labels for all domains
}
