% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{make_acs_itt_set}
\alias{make_acs_itt_set}
\title{Create the ACS-ITT set with relevant variables}
\usage{
make_acs_itt_set(dat)
}
\arguments{
\item{dat}{A dataset}
}
\value{
The ACS-ITT set
}
\description{
Creates the ACS-ITT analysis set for the primary outcome.
May need some tweaking for secondary outcomes.
}
