% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{make_avs_itt_set}
\alias{make_avs_itt_set}
\title{Create the AVS-ITT set with relevant variables}
\usage{
make_avs_itt_set(dat)
}
\arguments{
\item{dat}{A dataset}
}
\value{
The AVS-ITT set
}
\description{
Creates the AVS-ITT analysis set for the primary outcome.
May need some tweaking for secondary outcomes.
}
