% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{make_domC_design}
\alias{make_domC_design}
\title{make_domC_design}
\usage{
make_domC_design(dat, ctr = contr.equalprior)
}
\arguments{
\item{dat}{A dataset with factor variable "randC" for which NA means not randomised.}

\item{ctr}{Contrast, by defualt orthonormal}
}
\value{
Design matrix
}
\description{
Create design matrix where domain has an intercept which
indicates whether a participant was randomised to it or not
and (by default) uses orthonormal contrasts for available
interventions.

This keeps the "not randomised" group out of the overall
mean around which interventions are centred.
}
