% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{make_fas_itt_set}
\alias{make_fas_itt_set}
\title{Create the FAS-ITT set with relevant variables}
\usage{
make_fas_itt_set(dat)
}
\arguments{
\item{dat}{A dataset}
}
\value{
The FAS-ITT set
}
\description{
Creates the FAS-ITT analysis set for the primary outcome.
May need some tweaking for secondary outcomes.
}
