% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{make_intervention_table}
\alias{make_intervention_table}
\title{Interevention assignment table}
\usage{
make_intervention_table(dat)
}
\arguments{
\item{dat}{Dataset with `PO`, `AAssignment`, and `CAssignment`.}
}
\value{
Tibble of intervention assignment summary
}
\description{
Interevention assignment table
}
