% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{make_primary_model_data}
\alias{make_primary_model_data}
\title{Make primary outcome stan data}
\usage{
make_primary_model_data(
  dat,
  outcome = "PO",
  vars = c("inelgc3", "agegte60", "ctry"),
  beta_sd_int = 2.5,
  beta_sd_var = c(10, 2.5, 1, 1),
  beta_sd_trt = 1,
  ctr = contr.equalprior,
  includeA = TRUE,
  includeC = TRUE,
  intercept = TRUE,
  ...
)
}
\arguments{
\item{dat}{Dataset}

\item{vars}{Variables to include in model}

\item{beta_sd_int}{Prior SD on intercept}

\item{beta_sd_var}{Prior SD on vars}

\item{beta_sd_trt}{Prior SD on treatment terms}

\item{ctr}{Contrast for treatments}

\item{..}{Other args}
}
\value{
A list of data for use with Stan model
}
\description{
Create a dataset for use with Stan
}
