% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{odds_ratio_summary_table_rev}
\alias{odds_ratio_summary_table_rev}
\title{Create odds ratio summary table}
\usage{
odds_ratio_summary_table_rev(OR, format = "html", fn = NULL)
}
\arguments{
\item{OR}{Odds ratio RV}

\item{format}{Table format}

\item{fn}{Filename}
}
\value{
Table if fn is NULL, otherwise save the table
}
\description{
Create odds ratio summary table
}
