% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{plot_epoch_site_terms}
\alias{plot_epoch_site_terms}
\title{Plot site and epoch terms posterior summaries in one figure}
\usage{
plot_epoch_site_terms(rvs_epoch, rvs_site, region)
}
\arguments{
\item{rvs_epoch}{Epoch RVs}

\item{rvs_site}{Site RVs}

\item{region}{Site region groupings}
}
\value{
A ggplot
}
\description{
Plot site and epoch terms posterior summaries in one figure
}
