% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{plot_epoch_terms}
\alias{plot_epoch_terms}
\title{Plot epoch term posterior summaries}
\usage{
plot_epoch_terms(rvs_epoch)
}
\arguments{
\item{rvs_epoch}{Epoch RVs}
}
\value{
A ggplot
}
\description{
Plot epoch term posterior summaries
}
