% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{plot_site_terms}
\alias{plot_site_terms}
\title{Plot site term posterior summaries}
\usage{
plot_site_terms(rvs_site, region, xlab = "Change in log-odds")
}
\arguments{
\item{rvs_site}{Site RVs}

\item{region}{Site region groupings}
}
\value{
A ggplot
}
\description{
Plot site term posterior summaries
}
