% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mappings.R
\name{pp_lwmh_dose}
\alias{pp_lwmh_dose}
\title{pp_lwmh_dose}
\usage{
pp_lwmh_dose()
}
\description{
Note for therapeutic dose
Enoxaparin: PP_Dose is 1 x BAS_Weight twice per day, or 1.5 x BAS_Weight once
per day. PP_Dose is in milligrams (mg) at the frequency in the 'times' column.
Tinzaparin: PP_Dose is IU/kg - Note only 1 patient received Tinzaparin and
none received Dalteparin.
Common to protocol 3 (incl C3 arm) and protocol 5.
Matches original pp_lmwh_dose table that did not account
for creatinine levels at baseline and assumes creatinine clearance >30mL/min.
}
