% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mappings.R
\name{pp_lwmh_dose_creat_under}
\alias{pp_lwmh_dose_creat_under}
\title{pp_lwmh_dose_creat_under}
\usage{
pp_lwmh_dose_creat_under()
}
\description{
Protocol 5 Dosing table is not applicable to C3 (standard dose + aspirin)
Different means differs from creatinine over >30 dosing
}
