% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_cuts}
\alias{quantile_cuts}
\title{quantile_cuts}
\usage{
quantile_cuts(x, ss)
}
\arguments{
\item{x}{Numeric vector}

\item{ss}{Increasing sequence of probability cut-points}
}
\value{
A factor giving the quantiles according to `ss`
}
\description{
Convert a numeric vector into a categorical according to quantiles.
}
