% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_data.R
\name{read_all_raw_extracts}
\alias{read_all_raw_extracts}
\title{read_all_raw_extracts}
\usage{
read_all_raw_extracts(dir)
}
\arguments{
\item{dir}{Data directory}
}
\value{
Returns nothing, but adds all data extracts to global environment.
}
\description{
Read all the raw data extracts in `dir`.
}
