% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_data.R
\name{read_eligibility_file}
\alias{read_eligibility_file}
\title{read_eligibility_file}
\usage{
read_eligibility_file(fn)
}
\arguments{
\item{fn}{Filename}
}
\value{
tibble of data
}
\description{
read_eligibility_file
}
