% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_data.R
\name{read_raw_extracts}
\alias{read_raw_extracts}
\title{read_raw_extracts}
\usage{
read_raw_extracts(dir, fn)
}
\arguments{
\item{dir}{Data directory}

\item{fn}{The name of the extract as a string, or a vector of extract names}
}
\value{
Returns nothing, but adds the extracted data to the global environment.
Always reads the most recent data (based on ordering of file name).
}
\description{
Read the raw data for extract `fn` from `dir`.
}
