% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_cmdstanr_model}
\alias{save_cmdstanr_model}
\title{save_cmdstanr_model}
\usage{
save_cmdstanr_model(mod, fn)
}
\arguments{
\item{mod}{cmdstanr model object}

\item{fn}{File name}
}
\value{
Nothing, but saves the model object to outputs/models/<fn>
}
\description{
Save cmdstanr model
}
