% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_derived_dataset}
\alias{save_derived_dataset}
\title{save_derived_dataset}
\usage{
save_derived_dataset(dir, dat, fn)
}
\arguments{
\item{dir}{The derived data directory}

\item{dat}{The dataset}

\item{fn}{The file name which will be appended to `dir`.}
}
\value{
Nothing
}
\description{
Save a derived dataset to the appropriate directory (ANTICOAG_DATA)
}
