% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_tex_table}
\alias{save_tex_table}
\title{save_tex_table}
\usage{
save_tex_table(tab, fn)
}
\arguments{
\item{tab}{The latex table}

\item{fn}{The file name (without .tex extension)}
}
\value{
Nothing, but writes the table to outputs/tables/<fn>.tex
}
\description{
Save a latex table
}
