% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_activation}
\alias{site_activation}
\title{Study site activation data}
\format{
## `site_activation`

A data frame:
\describe{
  \item{state}{Site state}
  \item{site}{Site code}
  \item{health_service}{Health service}
  \item{active_v3}{Start date on protocol version 3}
  \item{active_v5}{Start date on protocol version 5}
  \item{country}{Country of site}
}
}
\usage{
site_activation
}
\description{
Data on when study sites activated each study protocol and what domains
where available at the site.
}
\keyword{datasets}
