% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_completeness.R
\name{summarise_completeness_data}
\alias{summarise_completeness_data}
\title{summarise_completeness_data}
\usage{
summarise_completeness_data(cp_dat)
}
\arguments{
\item{cp_dat}{Output from `generate_completeness_data`.}
}
\value{
A number of tibbles summarising data completeness which can be used
with kableExtra to generate tables fro interim reports.
}
\description{
Summarise data completeness
}
