% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{summarise_daily_data}
\alias{summarise_daily_data}
\title{summarise_daily_data}
\usage{
summarise_daily_data(dd)
}
\arguments{
\item{dd}{Daily data extract with additional variables as required.}
}
\value{
Dataset with one row per patient summarising their daily data.
}
\description{
Summarise the daily data.
Requires fields from other baseline extracts, so
assumes all necessary variables are included in `dd`.
}
