% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{summarise_posterior}
\alias{summarise_posterior}
\title{Summarise a posterior}
\usage{
summarise_posterior(dat, futval = 1.1)
}
\arguments{
\item{dat}{Dataset with column `Posterior` of type `rvar`.}

\item{futval}{Futility reference value}
}
\value{
A tibble of posterior summaries
}
\description{
Summarise a posterior
}
