% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{transmute_model_cols_grp_aus_nz}
\alias{transmute_model_cols_grp_aus_nz}
\title{transmute_model_cols_grp_aus_nz}
\usage{
transmute_model_cols_grp_aus_nz(dat)
}
\arguments{
\item{dat}{Dataset}
}
\value{
Transformed dataset with limited variables
}
\description{
Same as transmute_model_cols, but join Australia and new zealand
together as one "region" with nested sites.
}
